Revert "moves get user favorites strain ids in to favorites middleware component"
This reverts commit 28834bfb2dcedef93c7b00acf1b65bc22193b0b0.
